- creating gui_login.py, gui_tracker.py, utils.py - transformation login.py, main.py
